Remove documentation files and unused scripts
