+++ requirements.txt
@@ -1,7 +1,3 @@
-requests>=2.28.0
-beautifulsoup4>=4.11.0
-selenium>=4.8.0
-aiohttp>=3.8.0
-lxml>=4.9.0
-webdriver-manager>=3.8.0
-Pillow>=9.0.0
+requests
+beautifulsoup4
+selenium
