--- targets.txt
@@ -1,12 +1,0 @@
-192.168.1.1
-192.168.1.254
-192.168.0.1
-192.168.0.254
-10.0.0.1
-10.0.0.254
-172.16.0.1
-172.16.0.254
-192.168.100.1
-192.168.100.254
-10.1.1.1
-10.1.1.254
